pyproject.toml
[tool.poetry]
name = "dom"
version = "0.0.1"
description = ""
authors = ["theguywhobuildsthings <theguywhobuildsthings@test.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
qrcode = {extras = ["pil"], version = "^7.3.1"}
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
pydantic = "^1.10.2"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
self = "^2020.12.3"
poetry-dotenv-plugin = "^0.1.0"
sqlalchemy = "^1.4.44"
websockets = "^10.4"
redis = "^4.4.0"
pytest = "^7.2.0"
httpx = "^0.23.1"
pytest-asyncio = "^0.20.3"
factory-boy = "^3.2.1"
flake8 = "^6.0.0"
flake8-black = "^0.3.6"
black = "^22.12.0"
autopep8 = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
